refactor: 코드 리펙토링
