+++ src/main/java/com/genius/gitget/topic/facade/TopicFacade.java
@@ -1,4 +1,4 @@
-package com.genius.gitget.topic.serviceFacade;
+package com.genius.gitget.topic.facade;
 
 import com.genius.gitget.topic.dto.TopicCreateRequest;
 import com.genius.gitget.topic.dto.TopicDetailResponse;
